refactor session expiration TS to remove redundancy
