+++ app/views/session_timeout/_expire_session.html.erb
@@ -1,4 +1,4 @@
-<%= tag.div id: 'js-expire-session',
+<%= tag.div id: 'session-timeout-cntnr',
             data: {
               session_timeout_in: session_timeout_in,
               warning: warning,
