+++ app/views/session_timeout/_ping.html.erb
@@ -7,4 +7,4 @@
               frequency: frequency,
             } %>
 <%= render(partial: 'session_timeout/warning', locals: { modal_presenter: modal }) %>
-<%= javascript_packs_tag_once 'session-timeout-ping' %>
+<%= javascript_packs_tag_once 'session-expire-session' %>
